requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
rcon>=2.4.0
httpx>=0.27.0
pyinstaller>=6.0.0
